pizza to spring boot + liquibase
перевел проект пиццы на spring boot, добавил liquibase с initial.sql
