+++ .gitignore
@@ -1,40 +1,70 @@
+*.class
+*.log
+*.ctxt
+*.jar
+*.war
+*.ear
+*.nar
+*.zip
+*.tar.gz
+*.rar
+*.swp
+.DS_Store
+Thumbs.db
+
+.gradle/
+build/
+!gradle/wrapper/gradle-wrapper.jar
+
+!**/src/main/**/build/
+!**/src/test/**/build/
+
 target/
-!.mvn/wrapper/maven-wrapper.jar
 !**/src/main/**/target/
 !**/src/test/**/target/
 
-### IntelliJ IDEA ###
-.idea/modules.xml
-.idea/jarRepositories.xml
-.idea/compiler.xml
-.idea/libraries/
-*.iws
+.idea/
 *.iml
 *.ipr
+*.iws
+out/
 
-### Eclipse ###
-.apt_generated
+.settings/
+.project
 .classpath
 .factorypath
-.project
-.settings
 .springBeans
 .sts4-cache
+.apt_generated
+bin/
 
-### NetBeans ###
 /nbproject/private/
 /nbbuild/
 /dist/
 /nbdist/
 /.nb-gradle/
-build/
-!**/src/main/**/build/
-!**/src/test/**/build/
 
-### VS Code ###
 .vscode/
 
-### Mac OS ###
-.DS_Store
 /history/
 /log/
+
+dependency-reduced-pom.xml
+build.log
+
+*.liquibase
+
+*.h2.db
+
+*.orig
+*~
+*.bak
+
+/reports/
+/test-output/
+/coverage/
+
+application.properties
+application.yml
+
+*.versionsBackup
