+++ pizza-spring/src/main/java/ru/cleancode/pizzaspring/exceptions/OrderLimitExceededException.java
@@ -1,0 +1,9 @@
+package ru.cleancode.pizzaspring.exceptions;
+
+import java.text.MessageFormat;
+
+public class OrderLimitExceededException extends Throwable {
+    public OrderLimitExceededException(int size, int countLimitDishes) {
+        super(MessageFormat.format("Превышен лимит блюд на кухне. Заказано {0} при лимите {1}", size, countLimitDishes));
+    }
+}
