+++ pizza-spring/src/main/java/ru/cleancode/pizzaspring/exceptions/PizzaPriceException.java
@@ -1,0 +1,7 @@
+package ru.cleancode.pizzaspring.exceptions;
+
+public class PizzaPriceException extends Throwable {
+    public PizzaPriceException() {
+        super("Введена неверная стоимость пиццы");
+    }
+}
