+++ pizza-spring/src/main/java/ru/cleancode/pizzaspring/objects/Dish.java
@@ -1,0 +1,23 @@
+package ru.cleancode.pizzaspring.objects;
+
+import java.math.BigDecimal;
+
+/**
+ * Блюдо
+ */
+public interface Dish {
+
+    /**
+     * Наименование блюда
+     *
+     * @return наименование блюда
+     */
+    String getDishName();
+
+    /**
+     * Цена блюда
+     *
+     * @return Цена блюда
+     */
+    BigDecimal getDishPrice();
+}
