+++ pizza-spring/src/main/java/ru/cleancode/pizzaspring/objects/Ingredient.java
@@ -1,0 +1,21 @@
+package ru.cleancode.pizzaspring.objects;
+
+import lombok.Builder;
+import lombok.Data;
+
+/**
+ * Ингредиент
+ */
+@Builder
+@Data
+public class Ingredient {
+    /**
+     * Наименование ингредиента
+     */
+    private String name;
+
+    /**
+     * Вес ингредиента
+     */
+    private int weight;
+}
