+++ pizza-spring/src/main/java/ru/cleancode/pizzaspring/objects/PizzaSize.java
@@ -1,0 +1,11 @@
+package ru.cleancode.pizzaspring.objects;
+
+import lombok.RequiredArgsConstructor;
+
+/**
+ * Размеры пицц
+ */
+@RequiredArgsConstructor
+public enum PizzaSize {
+    SMALL, MEDIUM, LARGE
+}
