+++ pizza-spring/src/main/java/ru/cleancode/pizzaspring/objects/kitchens/Pizzeria.java
@@ -1,0 +1,13 @@
+package ru.cleancode.pizzaspring.objects.kitchens;
+
+/**
+ * Спецификация для пиццерии
+ */
+public interface Pizzeria {
+    /**
+     * Является ли пиццерией
+     *
+     * @return true/false
+     */
+    boolean isPizzeria();
+}
