+++ pizza-spring/src/main/java/ru/cleancode/pizzaspring/objects/pizzas/PizzaDecorator.java
@@ -1,0 +1,24 @@
+package ru.cleancode.pizzaspring.objects.pizzas;
+
+import lombok.Getter;
+import ru.cleancode.pizzaspring.objects.Ingredient;
+
+import java.util.List;
+
+/**
+ * Базовый декоратор для пиццы
+ */
+@Getter
+public abstract class PizzaDecorator extends Pizza {
+    protected final Pizza pizza;
+
+    protected PizzaDecorator(Pizza pizza) {
+        super(pizza.getName(), pizza.getSize(), pizza.getPrice());
+        this.pizza = pizza;
+        this.ingredients = pizza.getIngredients();
+    }
+
+    public List<Ingredient> getIngredients() {
+        return pizza.getIngredients();
+    }
+}
