+++ pizza-spring/src/main/resources/db/changelog/db.changelog-master.yaml
@@ -1,0 +1,4 @@
+databaseChangeLog:
+  - include:
+      file: sql/initial.sql
+      relativeToChangelogFile: true
