+++ pizza-spring/src/test/java/ru/cleancode/pizzaspring/PizzaSpringApplicationTests.java
@@ -1,0 +1,13 @@
+package ru.cleancode.pizzaspring;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class PizzaSpringApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
