--- pizza/build.gradle
@@ -1,19 +1,0 @@
-plugins {
-    id 'java'
-}
-
-group = 'ru.cleancode'
-version = 'unspecified'
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    testImplementation platform('org.junit:junit-bom:5.10.0')
-    testImplementation 'org.junit.jupiter:junit-jupiter'
-}
-
-test {
-    useJUnitPlatform()
-}
