--- pizza/src/main/java/ru/cleancode/PizzaProject/exceptions/KitchenTypeException.java
@@ -1,9 +1,0 @@
-package ru.cleancode.PizzaProject.exceptions;
-
-import java.text.MessageFormat;
-
-public class KitchenTypeException extends Throwable {
-    public KitchenTypeException(Class<?> Class) {
-        super(MessageFormat.format("Получен список блюд, которые невозможно выполнить в {0}", Class.getName()));
-    }
-}
