--- pizza/src/main/java/ru/cleancode/PizzaProject/exceptions/PizzaPriceException.java
@@ -1,7 +1,0 @@
-package ru.cleancode.PizzaProject.exceptions;
-
-public class PizzaPriceException extends Throwable {
-    public PizzaPriceException() {
-        super("Введена неверная стоимость пиццы");
-    }
-}
