--- pizza/src/main/java/ru/cleancode/PizzaProject/objects/PizzaSize.java
@@ -1,11 +1,0 @@
-package ru.cleancode.PizzaProject.objects;
-
-import lombok.RequiredArgsConstructor;
-
-/**
- * Размеры пицц
- */
-@RequiredArgsConstructor
-public enum PizzaSize {
-    SMALL, MEDIUM, LARGE
-}
