--- pizza/src/main/java/ru/cleancode/PizzaProject/objects/kitchens/Pizzeria.java
@@ -1,13 +1,0 @@
-package ru.cleancode.PizzaProject.objects.kitchens;
-
-/**
- * Спецификация для пиццерии
- */
-public interface Pizzeria {
-    /**
-     * Является ли пиццерией
-     *
-     * @return true/false
-     */
-    boolean isPizzeria();
-}
