--- pizza/src/main/java/ru/cleancode/PizzaProject/utility/LoggerFactory.java
@@ -1,10 +1,0 @@
-package ru.cleancode.PizzaProject.utility;
-
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-
-public class LoggerFactory {
-    public static Logger getLogger(Class<?> clazz) {
-        return LogManager.getLogger(clazz);
-    }
-}
