+++ settings.gradle
@@ -3,6 +3,6 @@
  */
 
 rootProject.name = 'cleancode'
-include(':calculator')
-include 'pizza'
-include 'managerRefactor'
+include ':calculator'
+include ':managerRefactor'
+include ':pizza-spring'
